Backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1.2
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-rosetta==0.10.1
django-solo==2.1.0
exceptiongroup==1.2.2
executing==2.1.0
getmac==0.9.5
gunicorn==23.0.0
idna==3.10
ipython==8.28.0
jedi==0.19.1
kombu==5.4.2
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
polib==1.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
